server/requirements.txt
flask
pyyaml
python-ldap
